Fix structure for uboot for our mx4 machines
I think it is possible to make it even more pretty but for now it should work the same as it did on the 2016.11.hm fork of uboot-toradex. It simply copy the files it need to the work dir. I use the similar structure as the kernel with machinoverides when it is needed so it should be easier to make changes to the diffrent machine or "machine groups" in the future. this will also make our forked toradex-uboot not needed anymore.